show user menu list in mobile menu when logged in
